# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Item Test API",
        "description": "API untuk mengelola data dokter phc",
        "version": "1.0"
    },
    "paths": {
        "/api/itemtest": {
            "get": {
                "tags": [
                    "Item Test"
                ],
                "summary": "Get all Item Tests or search by ID or query (Urine filter)",
                "operationId": "40c8af15f508d56a05797a33d267f350",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Search by exact ID (integer match)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search by text query (LIKE match, 'urine' filter applied)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of Item Tests (filtered by ID or Urine query)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ItemTest"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Item Test"
                ],
                "summary": "Tambah data item baru",
                "description": "Menambahkan item baru ke dalam database.",
                "operationId": "24aa2a8a1691ea73344effeee35eaa80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ItemTestCode",
                                    "ItemTestName"
                                ],
                                "properties": {
                                    "ItemTestCode": {
                                        "type": "string",
                                        "example": "ITC-001"
                                    },
                                    "ItemTestName": {
                                        "type": "string",
                                        "example": "Test Hemoglobin"
                                    },
                                    "Group": {
                                        "type": "string",
                                        "example": "Hematologi"
                                    },
                                    "SubGroup": {
                                        "type": "string",
                                        "example": "Darah"
                                    },
                                    "Descriptions": {
                                        "type": "string",
                                        "example": "Tes kadar hemoglobin dalam darah"
                                    },
                                    "CreateBy": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemTest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/itemtest/{ID}": {
            "put": {
                "tags": [
                    "Item Test"
                ],
                "summary": "Update an existing Item Test",
                "description": "Memperbarui item test berdasarkan ID.",
                "operationId": "0689c15d32d597ce2280e4a6be1f6e6a",
                "parameters": [
                    {
                        "name": "ID",
                        "in": "path",
                        "description": "ID dari Item Test yang akan diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ItemTestCode",
                                    "ItemTestName"
                                ],
                                "properties": {
                                    "ItemTestName": {
                                        "type": "string",
                                        "example": "Test Hemoglobin"
                                    },
                                    "Group": {
                                        "type": "string",
                                        "example": "Hematologi"
                                    },
                                    "SubGroup": {
                                        "type": "string",
                                        "example": "Darah"
                                    },
                                    "Descriptions": {
                                        "type": "string",
                                        "example": "Tes kadar hemoglobin dalam darah"
                                    },
                                    "LastModifiedBy": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemTest"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item Test tidak ditemukan"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ItemTest": {
                "properties": {
                    "ItemTestCode": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "ItemTestName": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "Group": {
                        "type": "string",
                        "maxLength": 30,
                        "nullable": true
                    },
                    "SubGroup": {
                        "type": "string",
                        "maxLength": 30,
                        "nullable": true
                    },
                    "Descriptions": {
                        "type": "string",
                        "maxLength": 250,
                        "nullable": true
                    },
                    "CreateDate": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "CreateBy": {
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}